refactor: move token validation to common module
