--- application/src/main/kotlin/com/dobby/token/SecurityManager.kt
@@ -1,9 +1,0 @@
-package com.dobby.token
-
-interface SecurityManager {
-    fun generateAccessToken(memberId: String): String
-    fun generateRefreshToken(memberId: String): String
-    fun parseAuthentication(token: String): SecurityUser
-    fun getMemberIdFromRefreshToken(refreshToken: String): String
-    fun getMemberIdFromAccessToken(accessToken: String): String
-}
