+++ common/src/main/kotlin/com/dobby/security/SecurityManager.kt
@@ -1,0 +1,5 @@
+package com.dobby.security
+
+interface SecurityManager {
+    fun parseAuthentication(token: String): SecurityUser
+}
