+++ infrastructure/src/main/kotlin/com/dobby/config/SecurityConfig.kt
@@ -2,7 +2,7 @@ package com.dobby.config
 
 import com.dobby.security.token.JwtTokenManager
 import com.dobby.security.token.JwtTokenProvider
-import com.dobby.token.SecurityManager
+import com.dobby.security.SecurityManager
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
