+++ presentation/src/main/kotlin/com/dobby/api/config/filter/JwtOptionalAuthenticationFilter.kt
@@ -1,7 +1,7 @@
 package com.dobby.api.config.filter
 
-import com.dobby.token.SecurityManager
-import com.dobby.token.SecurityUser
+import com.dobby.security.SecurityManager
+import com.dobby.security.SecurityUser
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
